This pull request migrates the MECOMP project from using the tarpc RPC
framework to using gRPC via the tonic library, and updates the codebase
and dependencies accordingly. It also introduces a new crate,
`mecomp-prost`, for handling gRPC/protobuf types, and updates the CLI
and printing logic to use the new types and client interfaces.
Additionally, the PR improves test coverage command configuration and
optimizes build profiles for several dependencies.

The most important changes are:

### Migration from tarpc to gRPC/tonic

* Replaces all tarpc-based client and type usage in the CLI and related
modules with gRPC-based equivalents from the new `mecomp-prost` crate,
including `MusicPlayerClient` and data models. Updates all relevant
function signatures and usage patterns to match tonic/gRPC conventions.
[[1]](diffhunk://#diff-dc6952fe11603d2980ff27bcba97d6499f87dbb6f7813f015ae5cc62c2b2c7b0L6-R6)
[[2]](diffhunk://#diff-dc6952fe11603d2980ff27bcba97d6499f87dbb6f7813f015ae5cc62c2b2c7b0L35-R50)
[[3]](diffhunk://#diff-dc6952fe11603d2980ff27bcba97d6499f87dbb6f7813f015ae5cc62c2b2c7b0L65-R79)
[[4]](diffhunk://#diff-dc6952fe11603d2980ff27bcba97d6499f87dbb6f7813f015ae5cc62c2b2c7b0L91-R123)
[[5]](diffhunk://#diff-dc6952fe11603d2980ff27bcba97d6499f87dbb6f7813f015ae5cc62c2b2c7b0L137-R145)
[[6]](diffhunk://#diff-dc6952fe11603d2980ff27bcba97d6499f87dbb6f7813f015ae5cc62c2b2c7b0L160-R189)
[[7]](diffhunk://#diff-16866ef3d6a2b0ccbe6cd4fd363a3def2e1261449d6c7f4e1d57ac3c7b6a0068L22-R22)
[[8]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L3-R6)
[[9]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L102-R94)
[[10]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L128-R120)
[[11]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L167-R162)
[[12]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L176-R171)
[[13]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L186-R181)
[[14]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L197-R190)
[[15]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L207-R203)
* Removes tarpc from dependencies and configuration throughout the
workspace, including `Cargo.toml` and CLI-specific manifests.
[[1]](diffhunk://#diff-2e9d962a08321605940b5a657135052fbcef87b5e360662bb527c96d9a615542L150-R167)
[[2]](diffhunk://#diff-6057d83fb675b27e99f187fa354690b6597018c390cbd50eac854a56233d2addL32-R44)

### Introduction of `mecomp-prost` crate

* Adds the `mecomp-prost` crate to the workspace and as a dependency for
CLI and other crates, ensuring all protobuf-generated types and gRPC
clients are available for use.
[[1]](diffhunk://#diff-2e9d962a08321605940b5a657135052fbcef87b5e360662bb527c96d9a615542R13)
[[2]](diffhunk://#diff-2e9d962a08321605940b5a657135052fbcef87b5e360662bb527c96d9a615542R183)
[[3]](diffhunk://#diff-11e9ec4ce661f86495a00e3d45cf1f2def24253a9fc1f4c67f7c6536b87fb7c2R61)
[[4]](diffhunk://#diff-6057d83fb675b27e99f187fa354690b6597018c390cbd50eac854a56233d2addL32-R44)

### Dependency and build optimization

* Adds or updates several dependencies to optimize for test performance
(e.g., `cpal`, `rubato`, `tokio`, `tokio-stream`, `tonic`, `tower`,
`hyper`), and updates the `coverage` command to use
`--implicit-test-threads` for more reliable test coverage reporting.
[[1]](diffhunk://#diff-2e9d962a08321605940b5a657135052fbcef87b5e360662bb527c96d9a615542R80-R89)
[[2]](diffhunk://#diff-2e9d962a08321605940b5a657135052fbcef87b5e360662bb527c96d9a615542L150-R167)
[[3]](diffhunk://#diff-2e9d962a08321605940b5a657135052fbcef87b5e360662bb527c96d9a615542R183)
[[4]](diffhunk://#diff-9a4f3e4537ebb7474452d131b0d969d89a51286f4269aac5ef268e712be17268L4-R4)
[[5]](diffhunk://#diff-9626f7f6dcceefff6887f43f3959ca4ddd84e7c3466222d3dae7c753452ac0c3R53)

### CLI and data model updates

* Updates CLI command argument types for limits and indices from `usize`
to `u64` for better compatibility with gRPC/protobuf types.
[[1]](diffhunk://#diff-16866ef3d6a2b0ccbe6cd4fd363a3def2e1261449d6c7f4e1d57ac3c7b6a0068L76-R75)
[[2]](diffhunk://#diff-16866ef3d6a2b0ccbe6cd4fd363a3def2e1261449d6c7f4e1d57ac3c7b6a0068L358-R366)
* Updates printing and completion logic to use new gRPC/protobuf types
and field names, including changes to how IDs, artists, and durations
are handled.
[[1]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L102-R94)
[[2]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L128-R120)
[[3]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L167-R162)
[[4]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L176-R171)
[[5]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L186-R181)
[[6]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L197-R190)
[[7]](diffhunk://#diff-ae49a72e8c477eb10f56374a592f4ba07780d3033fd75f8983b48ca2fbdc98b9L207-R203)
[[8]](diffhunk://#diff-dc6952fe11603d2980ff27bcba97d6499f87dbb6f7813f015ae5cc62c2b2c7b0L65-R79)
[[9]](diffhunk://#diff-dc6952fe11603d2980ff27bcba97d6499f87dbb6f7813f015ae5cc62c2b2c7b0L91-R123)
[[10]](diffhunk://#diff-dc6952fe11603d2980ff27bcba97d6499f87dbb6f7813f015ae5cc62c2b2c7b0L137-R145)
[[11]](diffhunk://#diff-dc6952fe11603d2980ff27bcba97d6499f87dbb6f7813f015ae5cc62c2b2c7b0L160-R189)

### Documentation update

* Updates the project README to reflect the migration from tarpc to
gRPC/tonic for inter-process communication.

+++ storage/src/db/queries/generic.rs
@@ -125,12 +125,12 @@ pub fn read_related_in(
 /// Struct to assist deserializing the results of the count queries
 #[derive(Debug, serde::Deserialize, PartialEq, Eq, Clone, Copy)]
 pub struct Count {
-    count: usize,
+    count: u64,
 }
 
 impl Count {
     #[cfg(test)]
-    pub const fn new(count: usize) -> Self {
+    pub const fn new(count: u64) -> Self {
         Self { count }
     }
 
@@ -142,7 +142,7 @@ impl Count {
     pub async fn count<C: surrealdb::Connection>(
         db: &surrealdb::Surreal<C>,
         table: &str,
-    ) -> Result<usize, Error> {
+    ) -> Result<u64, Error> {
         let result: Option<Self> = db.query(count(table)).await?.take(0)?;
         Ok(result.map_or_else(
             || {
@@ -162,7 +162,7 @@ impl Count {
         db: &surrealdb::Surreal<C>,
         table: &str,
         relation: &str,
-    ) -> Result<usize, Error> {
+    ) -> Result<u64, Error> {
         let result: Option<Self> = db.query(count_orphaned(table, relation)).await?.take(0)?;
         Ok(result.map_or_else(
             || {
@@ -180,7 +180,7 @@ impl Count {
         table: &str,
         relation1: &str,
         relation2: &str,
-    ) -> Result<usize, Error> {
+    ) -> Result<u64, Error> {
         let result: Option<Self> = db
             .query(count_orphaned_both(table, relation1, relation2))
             .await?
